adding some javadocs
