--- src/main/java/ca/unknown/bot/use_cases/timer/TimerReply.java
@@ -1,4 +1,0 @@
-package ca.unknown.bot.use_cases.timer;
-
-public class TimerReply {
-}
